feat: unify share modal into reusable Stimulus controller
- Create share_modal_controller.js with fetch-and-render pattern
- Remove legacy turbo:load based share_modal.js module
- Update org_chart_controller.js to delegate share modal to new controller
- Update presence_controller.js to use share-modal#open action instead of redirect
- Update all share buttons to use data-action="click->share-modal#open"
- Add share-modal controller to creatives page, org chart, and comments popup
- Remove server-side rendered share modal from creatives/index.html.erb
- Share modal now works consistently across creative page, org chart, and chat
- Fixes bug where chat '+' button redirected instead of opening modal
