+++ engines/collavre/app/javascript/controllers/org_chart_controller.js
@@ -1,7 +1,7 @@
 import { Controller } from "@hotwired/stimulus"
 
 export default class extends Controller {
-  static targets = [ "shareContainer", "toggleAllBtn" ]
+  static targets = [ "toggleAllBtn" ]
 
   toggleAll() {
     const details = this.element.querySelectorAll("details.org-chart-group")
@@ -43,86 +43,4 @@ export default class extends Controller {
     })
   }
 
-  openShareModal(event) {
-    const btn = event.currentTarget
-    const sharesUrl = btn.dataset.sharesUrl
-
-    fetch(sharesUrl, {
-      headers: { "Accept": "text/html" }
-    })
-      .then(r => r.text())
-      .then(html => {
-        this.shareContainerTarget.innerHTML = html
-        this.#initializeModal()
-      })
-  }
-
-  closeShareModal() {
-    const modal = document.getElementById("share-creative-modal")
-    if (modal) {
-      modal.style.display = "none"
-      document.body.classList.remove("no-scroll")
-    }
-    this.shareContainerTarget.innerHTML = ""
-  }
-
-  // Private
-
-  #initializeModal() {
-    const modal = document.getElementById("share-creative-modal")
-    if (!modal) return
-
-    modal.style.display = "flex"
-    document.body.classList.add("no-scroll")
-
-    // Bind close button
-    const closeBtn = document.getElementById("close-share-modal")
-    if (closeBtn) {
-      closeBtn.onclick = () => this.closeShareModal()
-    }
-
-    // Close on backdrop click
-    modal.onclick = (e) => {
-      if (e.target === modal) this.closeShareModal()
-    }
-
-    // Initialize invite link button
-    this.#initializeInviteLink()
-  }
-
-  #initializeInviteLink() {
-    const inviteLinkBtn = document.getElementById("creative-invite-link")
-    if (!inviteLinkBtn) return
-
-    inviteLinkBtn.onclick = () => {
-      const creativeId = inviteLinkBtn.dataset.creativeId
-      const permissionSelect = document.getElementById("share-permission")
-      const permission = permissionSelect ? permissionSelect.value : "read"
-      const permissionLabel = permissionSelect ? permissionSelect.options[permissionSelect.selectedIndex].text : ""
-      const noAccessMessage = inviteLinkBtn.dataset.noAccessMessage
-      const copiedTemplate = inviteLinkBtn.dataset.copiedTemplate
-
-      if (permission === "no_access") {
-        alert(noAccessMessage)
-        return
-      }
-
-      fetch("/invite", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "X-CSRF-Token": document.querySelector("meta[name='csrf-token']")?.content
-        },
-        body: JSON.stringify({ creative_id: creativeId, permission })
-      })
-        .then(r => r.json())
-        .then(data => {
-          if (navigator.clipboard && navigator.clipboard.writeText) {
-            navigator.clipboard.writeText(data.url).then(() => {
-              alert(copiedTemplate.replace("__PERMISSION__", permissionLabel))
-            })
-          }
-        })
-    }
-  }
 }
