+++ engines/collavre/app/views/collavre/creatives/_share_button.html.erb
@@ -1,3 +1,6 @@
-<button id="share-creative-btn" class="btn btn-primary desktop-only">
+<% current_creative = @parent_creative || @creative %>
+<button id="share-creative-btn" class="btn btn-primary desktop-only"
+        data-action="click->share-modal#open"
+        data-share-modal-url-param="<%= collavre.creative_creative_shares_path(current_creative) %>">
   <span aria-hidden="true"><%= svg_tag 'share.svg', class: 'icon-up', width: 22, height: 20 %></span>
 </button>
