+++ engines/collavre/app/views/collavre/users/_org_chart.html.erb
@@ -1,6 +1,6 @@
-<div class="org-chart" data-controller="org-chart">
+<div class="org-chart" data-controller="org-chart share-modal">
   <%# Container for dynamically loaded share modal %>
-  <div data-org-chart-target="shareContainer"></div>
+  <div data-share-modal-target="container"></div>
 
   <% if org_chart_roots.present? %>
     <div class="org-chart-toolbar">
